backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
Pillow
opencv-python
ultralytics
accelerate
